💡 chore    : 기경능 추가 없이 코드 수정 (오타, 주석 등) ♻️ refactor: core presigned_url 시리얼라이저 리팩터링
- request.py + response.py -> presigned_url.py
